+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add `Pledge.resolve` and `Pledge.reject`
 * `Function.parallel` and `Function.series` now accept an array of Pledges/Promises
 * Add `Pledge.all` support & some more rejection fixes
+* Add `Pledge.race` and make it so the state can't change
 
 ## 0.5.5 (2018-06-18)